Fix SD detect pin usage
SD detect pin was hardcoded in sd_conf.h while API provide a way to specify it.
This requires to provide the pin number instead of the GPIO_PIN_x pin and GPIOX port.
That's why sd_conf.h has been removed from the core.
Moreove several definition for the BSP are common or only dependent to SD
instance.
SD instance could be SDMMC1 or SDIO depending of the STM32 series.
Only the SD detect definition should be redefined thanks the variant.h
if the board has one.

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>